+++ docs/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is (read: strives to be) based on Keep a Changelog.
 
 ---
 
+## [0.1.15] – 2026-01-23
+### Changed:
+- dwroshak-access bumped to 0.1.23
+
+### Added:
+- import and export funtions
+
+
+---
+
 ## [0.1.14] – 2026-01-23
 ### Changed:
 - dwroshak-access bumped to 0.1.21